Adding Layout and reorganize folder arhitecture
