--- src/components/Layout/index.ts
@@ -1,21 +1,0 @@
-import { Footer } from "src/components/Layout/Footer/index"
-import { Header } from "src/components/Layout/Header/index"
-import { Login } from "src/components/Layout/Login/index"
-import { Card } from "src/components/Layout/Card/index"
-import { CardSection } from "src/components/Layout/Card/section-card"
-import { Spinner } from "./Spinner/spinner"
-import { Profile } from "./subscribeInfo"
-import { VideoLine } from "./VideosLine"
-import { SetingsMenu } from "./SetingsMenu"
-
-export {
-  Footer,
-  Header,
-  Login,
-  Card,
-  CardSection,
-  Spinner,
-  Profile,
-  VideoLine,
-  SetingsMenu,
-}
