+++ src/components/index.ts
@@ -1,2 +1,23 @@
-import { Footer, Header, Login, Card, CardSection, Spinner, Profile, VideoLine } from "src/components/Layout/index"
-export { Footer, Header, Login, Card, CardSection, Spinner, Profile, VideoLine }
+import { Footer } from "src/components/Footer/index"
+import { Header } from "src/components/Header/index"
+import { Login } from "src/components/Login/index"
+import { Card } from "src/components/Card/index"
+import { CardSection } from "src/components/Card/section-card"
+import { Spinner } from "./Spinner/spinner"
+import { Profile } from "./subscribeInfo"
+import { VideoLine } from "./VideosLine"
+import { SetingsMenu } from "./SetingsMenu"
+import { Layout } from "./Layout"
+
+export {
+  Footer,
+  Header,
+  Login,
+  Card,
+  CardSection,
+  Spinner,
+  Profile,
+  VideoLine,
+  SetingsMenu,
+  Layout,
+}
