+++ src/components/subscribeInfo/index.tsx
