+++ src/pages/index.tsx
@@ -1,6 +1,7 @@
 import { magic } from "lib/magic-client"
 import { useEffect, useState } from "react"
 import { Login, Header, Footer, Spinner, VideoLine } from "src/components/index"
+import { Layout } from "src/components/index"
 
 const Homepage = () => {
   const [auth, setAuth] = useState<string | boolean>("loading")
@@ -18,11 +19,11 @@ const Homepage = () => {
   }, [auth])
 
   let homePage = (
-    <div className="flex flex-col bg-slate-50 min-h-screen w-full">
+    <Layout>
       <Header />
       <VideoLine />
       <Footer />
-    </div>
+    </Layout>
   )
   let LoginUser = (
     <div>
